+++ climate-data-ca/resources/js/analyze-functions.js
@@ -402,7 +402,32 @@
         $('.input-variable .input-item').click(function (e) {
 
             if (!$(this).hasClass('selected')) {
-                var var_content = JSON.parse($(this).closest('.input-variable').attr('data-content'))
+
+//         input_id = $(this).find('input').
+
+                let var_content = JSON.parse($(this).closest('.input-variable').attr('data-content'));
+                let var_frequencies = $(this).closest('.input-variable').attr('data-frequencies');
+
+                if (var_frequencies != '') {
+                    let frequencies = var_frequencies.split('|');
+                    $('input[name=freq]').each(function(i) {
+                       if (frequencies.includes($(this).val())) {
+                           $(this).attr('disabled', false);
+                        } else {
+                           let input_row = $(this).closest('.input-row');
+                           input_row.removeClass('checked');
+                           $(this).prop('checked', false);
+                           input_row.find('.form-icon').removeClass().addClass(radio_icon_off);
+                           $(this).attr('disabled', true);
+                       }
+                    });
+                } else {
+                    $('input[name=freq]').each(function(i) {
+                        $(this).attr('disabled', false);
+                    });
+                }
+
+
                 var new_html = ''
                 new_html += '<h4>' + $(this).find('label').text() + '</h4>'
                 new_html += '<div class="form-inline">'
@@ -467,6 +492,19 @@
                         });
 
                         new_html += '</select>';
+                    } else if (this_input['type'] == 'mm_dd') {
+
+                        new_html += '<input ';
+                        new_html += 'type="text" ';
+                        new_html += 'placeholder="MM-DD" ';
+                        new_html += 'class="form-control bg-transparent border-white mx-2 text-white" ';
+                        new_html += 'size="5" ';
+                        new_html += 'autocomplete="off" ';
+                        new_html += 'data-units="' + this_input['units'] + '" ';
+                        new_html += 'data-optional="' + this_input['optional'] + '" ';
+                        new_html += 'name="' + this_input['id'] + '" ';
+                        new_html += '>'
+
                     }
 
                 }
@@ -1150,14 +1188,17 @@
 
             $('#analyze-detail').find(':input').each(function () {
 
-                var this_name = $(this).attr('name')
+                var this_name = $(this).attr('name');
+                if ($(this).attr('data-optional') == undefined ||
+                    $(this).attr('data-optional') == true || $(this).val() != '') {
 
-                form_thresholds[this_name] = $(this).val()
+                    form_thresholds[this_name] = $(this).val();
 
-                if ($(this).val() != '' && $(this).attr('data-units') !== undefined && $(this).attr('data-units') != '') {
-                    form_thresholds[this_name] += ' ' + $(this).attr('data-units');
+                    if ($(this).val() != '' && $(this).attr('data-units') !== undefined && $(this).attr('data-units') != '') {
+                        form_thresholds[this_name] += ' ' + $(this).attr('data-units');
+                    }
                 }
-            })
+            });
 
             // cycle through the object and check for empty values
 
@@ -1165,7 +1206,6 @@
                 if (form_thresholds[key] == '') {
                     is_valid = false
                     thresholds_have_val = false
-                    console.log(key)
                 }
             }